backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
torch
transformers
accelerate
Pillow
python-dateutil
ffmpeg-python
